week-2_2/requirements.txt
apache-airflow-providers-amazon
psycopg2-binary
pyarrow
pandas
SQLAlchemy